Improvements to README and notebook
The README has been updated with more information of running the notebook.

An example of how to load csv style files has been added.  The option to display the RMS value of the circle fit in now included.
New code has been added to display the Arai plot of the data and new code has been added to return the ziggie values of all fits with n of 3 or more from the data.